fix(volume): persister le volume à la virgule — la zone ne se rallume plus muette
`zones.volume` était un INTEGER 0..100. L'écriture arrondissait
`(v * 100).round()`, ce qui coûte jusqu'à 3 dB vers −37 dB et, sous
0,005 linéaire — soit 20·log10(0,005) = −46,0205999133 dB EXACTEMENT —
tombait à 0 : après un redémarrage la zone se rallumait MUETTE,
indiscernable d'un mute volontaire.

Le seuil n'est pas « environ −46 dB » : c'est le point où `f64::round`
bascule de 1 à 0. `round` arrondit à l'opposé de zéro, donc l'entier vaut
0 pour tout v strictement inférieur à 0,005.

La colonne passe à virgule, l'ÉCHELLE NE BOUGE PAS : elle reste en
pour-cent 0..100, donc aucune ligne existante n'a besoin d'être
convertie et une migration à moitié appliquée ne peut multiplier ou
diviser le volume de personne par 100.

* `Zone.volume` : i32 → f64 ; `update_volume(id, f64)`, sans arrondi.
* Relecture par `as_f64()` — accepte Int, Real ET Text, les trois formes
  que les deux moteurs peuvent rendre.
* SQLite neuf : `volume REAL`. Pas de migration SQLite : le typage
  dynamique fait que l'affinité INTEGER ne convertit une décimale que si
  c'est SANS PERTE, et un test le prouve sur une table déclarée à
  l'ancienne au lieu de le supposer.
* PostgreSQL : migration 047 (idempotente, couvre integer / text /
  double precision), schéma 001 et cible de la 013 alignés. Sans elle PG
  REFUSE un f64 dans une colonne integer.

Trois chemins qui restaient nus, fermés au passage :
* `poller.rs` ×3 : `(vol * 100.0) as i32` TRONQUAIT — seuil −40 dB sur
  le volume adopté du renderer ;
* `config_backup.rs` ×2 : `as_i64()` rendait None sur un volume à
  virgule, et la restauration reposait 50 % en travers du réglage ;
* `zone_manager.rs` : idem à l'activation d'un profil de zone, le volume
  était silencieusement ignoré.

Refs #2886

+++ tune-core/migrations/postgres/001_initial_schema.sql
@@ -124,7 +124,8 @@ CREATE TABLE IF NOT EXISTS zones (
     name TEXT NOT NULL,
     output_type TEXT,
     output_device_id TEXT,
-    volume INTEGER DEFAULT 50,
+    -- #2886 : voir migration 047. DOUBLE PRECISION, echelle 0..100.
+    volume DOUBLE PRECISION DEFAULT 50,
     muted SMALLINT DEFAULT 0,
     online SMALLINT DEFAULT 1,
     gapless_enabled SMALLINT DEFAULT 1,
